Enabled previously disabled weights test
